Main Window: Move Spy class to utils
